# internal/adapter/http/handlers_list_transactions.go
package http

import (
	"errors"
	"net/http"
	"strconv"

	"github.com/formancehq/ledger/v3/internal/proto/commonpb"
)

// handleListTransactions handles GET /{ledgerName}/transactions to list a
// ledger's transactions. Query params:
//
//   - pageSize
//   - after=<txID>       cursor (exclusive)
//   - reverse=true       reverse the default order — see convention below
//   - startDate/endDate  RFC3339, filter on transaction timestamp.
//     Requires the builtin `TX_BUILTIN_INDEX_TIMESTAMP` index to be
//     enabled on the ledger via `CreateIndex`.
//   - filter             textual filterexpr grammar OR structured v2 JSON DSL
//     (EN-1511). A reference selection is expressed through it as the structured
//     `{"$match":{"reference":"<ref>"}}` (structured-only — the textual grammar
//     has no `reference` field); there is no separate `reference` alias.
//
// Ordering convention (mirrors `ctrl.Controller.ListTransactions`): the
// default (reverse=false) returns newest-first (descending transaction id);
// reverse=true returns oldest-first (ascending).
//
// Read-consistency options gap (tracked follow-up): the gRPC ListTransactions
// honours ReadOptions.checkpointId to pin a read to a
// specific applied index; this HTTP route deliberately does NOT expose them
// and serves a live, best-effort read of the current committed state. Clients
// that need a consistency-bounded / checkpoint-pinned read must use gRPC. This
// mirrors the same carve-out already made for the audit reads (EN-1481) and
// keeps EN-1472 scoped to "expose the reads over HTTP", not "full read-options
// parity". Same applies to the bucket reads (signing keys) below.
func (s *Server) handleListTransactions(w http.ResponseWriter, r *http.Request) {
	// The profile clock started in withQueryProfile, outside the scope guard, so
	// authentication as well as query-parameter parsing and filter compilation
	// land in PrepareDuration (EN-1859). drainCursor and the controller read the
	// same profile off this context.
	ctx := r.Context()
	profile := profileFromRequest(r)

	ledgerName, ok := requireLedgerName(w, r)
	if !ok {
		return
	}

	pageSize, ok := parsePageSize(w, r)
	if !ok {
		return
	}

	var afterTxID uint64

	if after := r.URL.Query().Get("after"); after != "" {
		parsed, err := strconv.ParseUint(after, 10, 64)
		if err != nil {
			writeBadRequest(w, "INVALID_REQUEST", errors.New("invalid after parameter"))

			return
		}

		afterTxID = parsed
	}

	reverse := r.URL.Query().Get("reverse") == "true"

	var filters []*commonpb.QueryFilter

	dateCond := &commonpb.UintCondition{}
	hasDateFilter := false

	if sd := r.URL.Query().Get("startDate"); sd != "" {
		v, ok := parseFilterDateMicros(w, "startDate", sd)
		if !ok {
			return
		}

		dateCond.Min = &v
		hasDateFilter = true
	}

	if ed := r.URL.Query().Get("endDate"); ed != "" {
		v, ok := parseFilterDateMicros(w, "endDate", ed)
		if !ok {
			return
		}

		dateCond.Max = &v
		dateCond.MaxExclusive = true
		hasDateFilter = true
	}

	if hasDateFilter {
		filters = append(filters, &commonpb.QueryFilter{
			Filter: &commonpb.QueryFilter_BuiltinUint{
				BuiltinUint: &commonpb.BuiltinUintCondition{
					Field: commonpb.TransactionBuiltinIndex_TX_BUILTIN_INDEX_TIMESTAMP,
					Cond:  dateCond,
				},
			},
		})
	}

	// The `filter` query parameter accepts either the textual filterexpr grammar
	// or the structured v2 JSON DSL (EN-1511); it is AND-combined with the
	// startDate/endDate timestamp range above.
	generic, ok := parseListFilter(w, r, commonpb.QueryTarget_QUERY_TARGET_TRANSACTIONS)
	if !ok {
		return
	}

	filters = append(filters, generic)

	filter := combineFilters(filters...)

	profile.EnterExecute()
	cursor, err := s.backend.ListTransactions(ctx, ledgerName, pageSize, afterTxID, filter, reverse)
	profile.LeaveExecute()

	if err != nil {
		handleError(w, r, err)

		return
	}

	transactions, ok := drainCursor(w, r, cursor)
	if !ok {
		return
	}

	finishProfile(w, r, profile)
	writeOKChecked(w, r, transactions)
}