# openapi.yml
openapi: 3.0.3
info:
  title: Ledger v3 API
  description: API for the Ledger v3, a distributed accounting ledger with Raft consensus
  version: 1.0.0
  contact:
    name: Formance
    url: https://formance.com
servers:
  - url: http://localhost:9000
    description: Local development server
security:
  - BearerAuth: []
  - {}
paths:
...
  /v3/{ledgerName}/indexes:
    get:
      summary: List indexes for a ledger
      description: |
        Streams the indexes registered on a specific ledger. Requires
        `ledger:LedgerRead` scope.
      operationId: listLedgerIndexes
      tags:
        - Indexes
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/LedgerName'
      responses:
        '200':
          description: Indexes retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      description: Index (protobuf JSON representation)
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Create an index on a ledger
      description: |
        Registers a new index on the ledger. The FSM begins the backfill on
        the next apply. The request body carries the target `IndexID` in
        canonical form (see `DELETE /v3/{ledgerName}/indexes/{canonicalId}`
        for the format). Requires `ledger:LedgerWrite` scope.
      operationId: createIndex
      tags:
        - Indexes
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/LedgerName'
        - name: Idempotency-Key
          in: header
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id:
                  type: string
                  description: |
                    Canonical form of the target IndexID. One of:
                      `tx_builtin:<TX_BUILTIN_INDEX_*>`
                      `log_builtin:<LOG_BUILTIN_INDEX_*>`
                      `account_builtin:<ACCT_BUILTIN_INDEX_*>`
                      `metadata:<TARGET_TYPE_*>:<key>`
                  example: metadata:TARGET_TYPE_ACCOUNT:color
      responses:
        '201':
          description: |
            Index creation queued. The body carries the canonical id of the
            created index (round-trip normalized), usable directly on the
            follow-up GET/DELETE `/v3/{ledgerName}/indexes/{canonicalId}` routes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateIndexResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT token issued by a trusted OIDC provider. Required when the server
        is started with `--auth-enabled`. Scopes: `ledger:read` (read operations),
        `ledger:write` (write operations), `ledger:admin` (cluster operations).
  parameters:
    LedgerName:
      name: ledgerName
      in: path
      required: true
      schema:
        type: string
      description: Name of the ledger
...
  responses:
...
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required (missing or invalid JWT token)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permissions (missing required scope)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict (e.g. idempotency key conflict, ledger deleted)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    CreateIndexResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - id
          properties:
            id:
              type: string
              description: |
                Canonical form of the created IndexID (round-trip normalized),
                e.g. `metadata:TARGET_TYPE_ACCOUNT:color` or
                `tx_builtin:TX_BUILTIN_INDEX_TIMESTAMP`.
              example: metadata:TARGET_TYPE_ACCOUNT:color
...
    ErrorResponse:
      type: object
      required:
        - errorCode
        - errorMessage
      properties:
        errorCode:
          type: string
          description: Error code identifier
        errorMessage:
          type: string
          description: Human-readable error message
...